Fix pulse logic, every pulse should be represented for reals this time
I think every pulse was being represented before but now I'm going by new logic to ensure it will not move on to the next pulse before one is expressed.

I have also added absolute caps to the microseconds a pulse can be expressed for before becoming 'noise', the values cannot go higher than the default settings and will not go lower than the floor of perception which may vary based on the individual.