Allow for higher pulse duration thresholds at lower resolution
Having pulse duration set too high will cause higher resolutions to not work because the contents of the audio file will be too big.

This is just for 50x50, to slow it down so the data is more interpretable (although it is still really fast)

As we determine the thresholds for pulse duration, it will be implemented into the resolution options for each pulse but for now I'm leaving it manually selectable.

The longer the pulse duration, the longer the audio output